% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_descriptive.R
\docType{data}
\name{AGE_BREAKS}
\alias{AGE_BREAKS}
\title{Age range breaks (in days) and labels}
\format{
An object of class \code{numeric} of length 8.
}
\usage{
AGE_BREAKS
}
\description{
Age range breaks (in days) and labels
}
\keyword{internal}
