% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_descriptive.R
\name{categorize_age}
\alias{categorize_age}
\title{Categorize age in days into age ranges}
\usage{
categorize_age(age_days)
}
\arguments{
\item{age_days}{Numeric vector of ages in days}
}
\value{
Factor of age ranges
}
\description{
Categorize age in days into age ranges
}
\keyword{internal}
