% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_descriptive.R
\name{categorize_substance_count}
\alias{categorize_substance_count}
\title{Categorize substance counts into ranges}
\usage{
categorize_substance_count(counts)
}
\arguments{
\item{counts}{Numeric vector of substance counts}
}
\value{
Factor of substance count categories
}
\description{
Categorize substance counts into ranges
}
\keyword{internal}
