% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_descriptive.R
\name{extract_year}
\alias{extract_year}
\title{Extract year from date fields with priority}
\usage{
extract_year(event_dt, init_fda_dt, fda_dt)
}
\arguments{
\item{event_dt}{Event date vector}

\item{init_fda_dt}{Initial FDA date vector}

\item{fda_dt}{FDA date vector}
}
\value{
Factor of years
}
\description{
Extract year from date fields with priority
}
\keyword{internal}
