% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowchart_generator.R
\name{flowchart_generator}
\alias{flowchart_generator}
\title{Generate a flowchart for a specified pregnancy algorithm
Creates a graphical flowchart summarizing inclusion and exclusion criteria
for pregnancy algorithms.}
\usage{
flowchart_generator(database = FAERS_version, pregnancy_flags_table, algorithm)
}
\arguments{
\item{database}{A FAERS@DiAna quarter (e.g., '25Q2') or the name of the
folder containing your input data. Defaults to \code{FAERS_version}.}

\item{pregnancy_flags_table}{A data frame returned by
\code{generate_pregnancy_flags()}.}

\item{algorithm}{Character string specifying the algorithm to display.
Allowed values: "UMC", "EMA", "SakaiHPPV", "SakaHS", "SakaiM".}
}
\value{
A ggplot object representing the flowchart.
}
\description{
Generate a flowchart for a specified pregnancy algorithm
Creates a graphical flowchart summarizing inclusion and exclusion criteria
for pregnancy algorithms.
}
\examples{
\dontrun{
flowchart_generator(
  database = FAERS_version,
  pregnancy_flags_table = flags,
  algorithm = "UMC"
)
}

}
