% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_Upset.R
\name{get_counts_for_upset}
\alias{get_counts_for_upset}
\title{Get Counts for UpSet Plot}
\usage{
get_counts_for_upset(subset, min_elements_in_intersection = 100)
}
\arguments{
\item{subset}{A data frame containing the data to be used for the UpSet plot.}

\item{min_elements_in_intersection}{An integer specifying the minimum number
of elements in an intersection to be included. Default is 100.}
}
\value{
A data frame with the counts of intersections and the concatenated
primary IDs for each intersection.
}
\description{
This function calculates the counts of intersections for an UpSet plot.
}
\details{
The function groups the data by all columns except \code{primaryid}, calculates
the count of primary IDs for each group, and concatenates the primary IDs into
a single string. The resulting data frame is sorted in descending order of the
counts.
}
\examples{
\dontrun{
# Example usage
get_counts_for_upset(subset, min_elements_in_intersection = 100)
}
}
