% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_descriptive.R
\name{load_pv_data}
\alias{load_pv_data}
\title{Load pharmacovigilance data from database or sample}
\usage{
load_pv_data(pids, database = "VigiBase", include_ther = FALSE)
}
\arguments{
\item{pids}{Character vector of primary IDs to load}

\item{database}{Database name ("sample", "VigiBase", or quarter like "24Q4")}

\item{include_ther}{Logical, whether to include therapy data}
}
\value{
List of data tables (Demo, Drug, Reac, Indi, Outc, Ther)
}
\description{
Load pharmacovigilance data from database or sample
}
\keyword{internal}
