% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_descriptive.R
\name{map_outcomes}
\alias{map_outcomes}
\title{Map outcome codes to readable labels}
\usage{
map_outcomes(outcome_codes)
}
\arguments{
\item{outcome_codes}{Character vector of outcome codes}
}
\value{
Character vector of outcome labels
}
\description{
Map outcome codes to readable labels
}
\keyword{internal}
