% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_descriptive.R
\name{map_reporter}
\alias{map_reporter}
\title{Map reporter occupation codes to readable labels}
\usage{
map_reporter(occp_codes)
}
\arguments{
\item{occp_codes}{Character vector of occupation codes}
}
\value{
Factor of reporter types
}
\description{
Map reporter occupation codes to readable labels
}
\keyword{internal}
