% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_descriptive.R
\name{map_sex}
\alias{map_sex}
\title{Map sex codes to readable labels}
\usage{
map_sex(sex_codes)
}
\arguments{
\item{sex_codes}{Character vector of sex codes (F, M, etc.)}
}
\value{
Character vector of sex labels
}
\description{
Map sex codes to readable labels
}
\keyword{internal}
