% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_descriptive.R
\name{map_submission}
\alias{map_submission}
\title{Map submission report codes to submission types}
\usage{
map_submission(rept_codes)
}
\arguments{
\item{rept_codes}{Character vector of report codes}
}
\value{
Character vector of submission types
}
\description{
Map submission report codes to submission types
}
\keyword{internal}
