% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_descriptive.R
\docType{data}
\name{OUTCOME_MAP}
\alias{OUTCOME_MAP}
\title{Outcome code mappings}
\format{
An object of class \code{character} of length 7.
}
\usage{
OUTCOME_MAP
}
\description{
Outcome code mappings
}
\keyword{internal}
