% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_Upset.R
\name{render_upset}
\alias{render_upset}
\title{Render UpSet Plot}
\usage{
render_upset(subset, min_elements_in_intersection = 100)
}
\arguments{
\item{subset}{A data frame containing the data to be used for the UpSet plot.}

\item{min_elements_in_intersection}{An integer specifying the minimum number
of elements in an intersection to be included. Default is 100.}
}
\value{
A ggplot object representing the UpSet plot.
}
\description{
This function renders an UpSet plot based on the provided subset of data.
}
\details{
The function processes the input data to handle missing values, converts
columns to numeric, and filters intersections based on the specified minimum
number of elements. It then orders the sets and creates the UpSet plot using
the ComplexUpset package.
}
\examples{
\dontrun{
# Example usage
render_upset(subset, min_elements_in_intersection = 100)
}

}
