% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_descriptive.R
\docType{data}
\name{REPORTER_MAP}
\alias{REPORTER_MAP}
\title{Reporter code mappings}
\format{
An object of class \code{character} of length 6.
}
\usage{
REPORTER_MAP
}
\description{
Reporter code mappings
}
\keyword{internal}
