% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_descriptive.R
\name{separate_table_column}
\alias{separate_table_column}
\title{Format and separate table output columns}
\usage{
separate_table_column(gt_table, col_index)
}
\arguments{
\item{gt_table}{Tibble containing gtsummary table}

\item{col_index}{Column index to separate}
}
\value{
Tibble with separated columns
}
\description{
Format and separate table output columns
}
\keyword{internal}
