% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_descriptive.R
\docType{data}
\name{SUBMISSION_MAP}
\alias{SUBMISSION_MAP}
\title{Submission code mappings}
\format{
An object of class \code{list} of length 3.
}
\usage{
SUBMISSION_MAP
}
\description{
Submission code mappings
}
\keyword{internal}
